Build my own custom emacs for redhat VMs
The main reason for this is the support for adding libvterm so that I
can have vterm working inside of emacs. The bams_build vm currently
does not build emacs with module support and I don't want to add that
while not officialy at KinetX anymore, so I build emacs again
privately.